[MODFEE-83] Extracted abstract parent for validation services
